+++ python_modules/dagster/dagster/core/types.py
@@ -392,3 +392,23 @@ Any = _DagsterAnyType()
 
 # TO DISCUSS: Consolidate with Dict?
 PythonDict = PythonObjectType('Dict', dict, type_attributes=DagsterTypeAttributes(is_builtin=True))
+
+
+def define_path_dict_field():
+    return Field(Dict({'path': Field(Path)}))
+
+
+class MaterializeableBuiltinScalarConfigSchema(ConfigurableSelectorFromDict):
+    def __init__(self, scalar_name):
+        # TODO: add pickle
+        super(
+            MaterializeableBuiltinScalarConfigSchema,
+            self,
+        ).__init__(fields={'json': define_path_dict_field()})
+
+        # Probably should come up with a better naming convention
+        self.name = scalar_name + '.MaterializationSchema'
+        self.description = ''
+
+    def iterate_types(self):
+        return []
